Provide ANSI-interpretted representations of test inputs/outputs (required splitting into two packages)
