+++ render/4space.go
@@ -1,5 +1,5 @@
 //go:build !2space && !6space && !8space
 
-package bean
+package render
 
 const indentSpaces = 4
