+++ test/test.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+go test ./1utilities.go $1
